Website Redesign: Improved UX and Additional Features
Updated index.html to be compatible with react 18
Refactored App.js to become more modular,  breaking down component and page in their own separate file.
Added Media for the Polish the Professional.
Added the Logos for the different projects.
Shortened the bio in the landing page.
Changed the main colors of the website.
Added a nav bar to fixed at the top of the page.
Refactored the CSS.
Added Footer, NavBar , PortfolioCard, ProjectCard, and PortfolioItem  Components.